Improve RRD4j implementation
- fix local timezone handling
- use rrd4j db pool
- always close() RrdDb files; otherwise the pool fills up and no more files can be opened
- apply coding best practices
  - apply Checkstyle
  - split interface and impl
